+++ twingate/internal/client/connector-tokens.go
@@ -4,16 +4,11 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/client/query"
 	"github.com/Twingate/terraform-provider-twingate/twingate/internal/model"
-	"github.com/twingate/go-graphql-client"
 )
 
 const connectorTokensResourceName = "connector tokens"
-
-type gqlConnectorTokens struct {
-	AccessToken  graphql.String
-	RefreshToken graphql.String
-}
 
 func (client *Client) VerifyConnectorTokens(ctx context.Context, refreshToken, accessToken string) error {
 	payload := map[string]string{
@@ -32,26 +27,17 @@ func (client *Client) VerifyConnectorTokens(ctx context.Context, refreshToken, a
 	return nil
 }
 
-type generateConnectorTokensQuery struct {
-	ConnectorGenerateTokens struct {
-		ConnectorTokens gqlConnectorTokens
-		OkError
-	} `graphql:"connectorGenerateTokens(connectorId: $connectorId)"`
-}
-
 func (client *Client) GenerateConnectorTokens(ctx context.Context, connectorID string) (*model.ConnectorTokens, error) {
 	variables := newVars(gqlID(connectorID, "connectorId"))
-	response := generateConnectorTokensQuery{}
+	response := query.GenerateConnectorTokens{}
 
 	err := client.GraphqlClient.NamedMutate(ctx, "generateConnectorTokens", &response, variables)
 	if err != nil {
 		return nil, NewAPIError(err, "generate", connectorTokensResourceName)
 	}
 
-	if !response.ConnectorGenerateTokens.Ok {
-		message := response.ConnectorGenerateTokens.Error
-
-		return nil, NewAPIErrorWithID(NewMutationError(message), "generate", connectorTokensResourceName, connectorID)
+	if !response.Ok {
+		return nil, NewAPIErrorWithID(NewMutationError(response.Error), "generate", connectorTokensResourceName, connectorID)
 	}
 
 	return response.ToModel(), nil
