+++ twingate/internal/client/connector.go
@@ -3,33 +3,12 @@ package client
 import (
 	"context"
 
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/client/query"
 	"github.com/Twingate/terraform-provider-twingate/twingate/internal/model"
 	"github.com/twingate/go-graphql-client"
 )
 
 const connectorResourceName = "connector"
-
-type gqlConnector struct {
-	IDName
-	RemoteNetwork struct {
-		ID graphql.ID
-	}
-}
-
-type createConnectorQuery struct {
-	ConnectorCreate struct {
-		Entity *gqlConnector
-		OkError
-	} `graphql:"connectorCreate(remoteNetworkId: $remoteNetworkId, name: $connectorName)"`
-}
-
-type ConnectorEdge struct {
-	Node *gqlConnector
-}
-
-type Connectors struct {
-	PaginatedResource[*ConnectorEdge]
-}
 
 func (client *Client) CreateConnector(ctx context.Context, remoteNetworkID, connectorName string) (*model.Connector, error) {
 	if remoteNetworkID == "" {
@@ -38,32 +17,25 @@ func (client *Client) CreateConnector(ctx context.Context, remoteNetworkID, conn
 
 	variables := newVars(
 		gqlID(remoteNetworkID, "remoteNetworkId"),
-		gqlNullableField(connectorName, "connectorName"),
+		gqlNullable(connectorName, "connectorName"),
 	)
 
-	response := createConnectorQuery{}
+	response := query.CreateConnector{}
 
 	err := client.GraphqlClient.NamedMutate(ctx, "createConnector", &response, variables)
 	if err != nil {
 		return nil, NewAPIErrorWithName(err, "create", connectorResourceName, connectorName)
 	}
 
-	if !response.ConnectorCreate.Ok {
-		return nil, NewAPIErrorWithName(NewMutationError(response.ConnectorCreate.Error), "create", connectorResourceName, connectorName)
+	if !response.Ok {
+		return nil, NewAPIErrorWithName(NewMutationError(response.Error), "create", connectorResourceName, connectorName)
 	}
 
-	if response.ConnectorCreate.Entity == nil {
+	if response.Entity == nil {
 		return nil, NewAPIErrorWithName(ErrGraphqlResultIsEmpty, "create", connectorResourceName, connectorName)
 	}
 
-	return response.ConnectorCreate.Entity.ToModel(), nil
-}
-
-type updateConnectorQuery struct {
-	ConnectorUpdate struct {
-		Entity *gqlConnector
-		OkError
-	} `graphql:"connectorUpdate(id: $connectorId, name: $connectorName)"`
+	return response.Entity.ToModel(), nil
 }
 
 func (client *Client) UpdateConnector(ctx context.Context, connectorID string, connectorName string) (*model.Connector, error) {
@@ -73,29 +45,25 @@ func (client *Client) UpdateConnector(ctx context.Context, connectorID string, c
 
 	variables := newVars(
 		gqlID(connectorID, "connectorId"),
-		gqlField(connectorName, "connectorName"),
+		gqlVar(connectorName, "connectorName"),
 	)
 
-	response := updateConnectorQuery{}
+	response := query.UpdateConnector{}
 
 	err := client.GraphqlClient.NamedMutate(ctx, "updateConnector", &response, variables)
 	if err != nil {
 		return nil, NewAPIErrorWithID(err, "update", connectorResourceName, connectorID)
 	}
 
-	if !response.ConnectorUpdate.Ok {
-		return nil, NewAPIErrorWithID(NewMutationError(response.ConnectorUpdate.Error), "update", connectorResourceName, connectorID)
+	if !response.Ok {
+		return nil, NewAPIErrorWithID(NewMutationError(response.Error), "update", connectorResourceName, connectorID)
 	}
 
-	if response.ConnectorUpdate.Entity == nil {
+	if response.Entity == nil {
 		return nil, NewAPIErrorWithID(ErrGraphqlResultIsEmpty, "update", connectorResourceName, connectorID)
 	}
 
-	return response.ConnectorUpdate.Entity.ToModel(), nil
-}
-
-type deleteConnectorQuery struct {
-	ConnectorDelete *OkError `graphql:"connectorDelete(id: $id)"`
+	return response.Entity.ToModel(), nil
 }
 
 func (client *Client) DeleteConnector(ctx context.Context, connectorID string) error {
@@ -104,22 +72,18 @@ func (client *Client) DeleteConnector(ctx context.Context, connectorID string) e
 	}
 
 	variables := newVars(gqlID(connectorID))
-	response := deleteConnectorQuery{}
+	response := query.DeleteConnector{}
 
 	err := client.GraphqlClient.NamedMutate(ctx, "deleteConnector", &response, variables)
 	if err != nil {
 		return NewAPIErrorWithID(err, "delete", connectorResourceName, connectorID)
 	}
 
-	if !response.ConnectorDelete.Ok {
-		return NewAPIErrorWithID(NewMutationError(response.ConnectorDelete.Error), "delete", connectorResourceName, connectorID)
+	if !response.Ok {
+		return NewAPIErrorWithID(NewMutationError(response.Error), "delete", connectorResourceName, connectorID)
 	}
 
 	return nil
-}
-
-type readConnectorQuery struct {
-	Connector *gqlConnector `graphql:"connector(id: $id)"`
 }
 
 func (client *Client) ReadConnector(ctx context.Context, connectorID string) (*model.Connector, error) {
@@ -128,7 +92,7 @@ func (client *Client) ReadConnector(ctx context.Context, connectorID string) (*m
 	}
 
 	variables := newVars(gqlID(connectorID))
-	response := readConnectorQuery{}
+	response := query.ReadConnector{}
 
 	err := client.GraphqlClient.NamedQuery(ctx, "readConnector", &response, variables)
 	if err != nil {
@@ -142,50 +106,39 @@ func (client *Client) ReadConnector(ctx context.Context, connectorID string) (*m
 	return response.ToModel(), nil
 }
 
-type readConnectorsQuery struct {
-	Connectors Connectors
-}
-
 func (client *Client) ReadConnectors(ctx context.Context) ([]*model.Connector, error) {
-	response := readConnectorsQuery{}
-
-	err := client.GraphqlClient.NamedQuery(ctx, "readConnectors", &response, nil)
-	if err != nil {
-		return nil, NewAPIErrorWithID(err, "read", connectorResourceName, "All")
-	}
-
-	if len(response.Connectors.Edges) == 0 {
-		return nil, NewAPIErrorWithID(ErrGraphqlResultIsEmpty, "read", connectorResourceName, "All")
-	}
-
-	err = response.Connectors.fetchPages(ctx, client.readConnectorsAfter, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return response.Connectors.ToModel(), nil
-}
-
-type readConnectorsAfter struct {
-	Connectors Connectors `graphql:"connectors(after: $connectorsEndCursor)"`
-}
-
-func (client *Client) readConnectorsAfter(ctx context.Context, variables map[string]interface{}, cursor graphql.String) (*PaginatedResource[*ConnectorEdge], error) {
-	if variables == nil {
-		variables = make(map[string]interface{})
-	}
-
-	variables["connectorsEndCursor"] = cursor
-	response := readConnectorsAfter{}
+	response := query.ReadConnectors{}
+	variables := newVars(gqlNullable("", query.CursorConnectors))
 
 	err := client.GraphqlClient.NamedQuery(ctx, "readConnectors", &response, variables)
 	if err != nil {
 		return nil, NewAPIErrorWithID(err, "read", connectorResourceName, "All")
 	}
 
-	if len(response.Connectors.Edges) == 0 {
+	if len(response.Edges) == 0 {
 		return nil, NewAPIErrorWithID(ErrGraphqlResultIsEmpty, "read", connectorResourceName, "All")
 	}
 
-	return &response.Connectors.PaginatedResource, nil
+	err = response.FetchPages(ctx, client.readConnectorsAfter, variables)
+	if err != nil {
+		return nil, err //nolint
+	}
+
+	return response.ToModel(), nil
+}
+
+func (client *Client) readConnectorsAfter(ctx context.Context, variables map[string]interface{}, cursor graphql.String) (*query.PaginatedResource[*query.ConnectorEdge], error) {
+	variables[query.CursorConnectors] = cursor
+	response := query.ReadConnectors{}
+
+	err := client.GraphqlClient.NamedQuery(ctx, "readConnectors", &response, variables)
+	if err != nil {
+		return nil, NewAPIErrorWithID(err, "read", connectorResourceName, "All")
+	}
+
+	if len(response.Edges) == 0 {
+		return nil, NewAPIErrorWithID(ErrGraphqlResultIsEmpty, "read", connectorResourceName, "All")
+	}
+
+	return &response.PaginatedResource, nil
 }
