+++ twingate/internal/client/query/connector-create.go
@@ -1,0 +1,10 @@
+package query
+
+type CreateConnector struct {
+	ConnectorEntityResponse `graphql:"connectorCreate(remoteNetworkId: $remoteNetworkId, name: $connectorName)"`
+}
+
+type ConnectorEntityResponse struct {
+	Entity *gqlConnector
+	OkError
+}
