+++ twingate/internal/client/query/connector-delete.go
@@ -1,0 +1,5 @@
+package query
+
+type DeleteConnector struct {
+	OkError `graphql:"connectorDelete(id: $id)"`
+}
