+++ twingate/internal/client/query/connector-update.go
@@ -1,0 +1,5 @@
+package query
+
+type UpdateConnector struct {
+	ConnectorEntityResponse `graphql:"connectorUpdate(id: $connectorId, name: $connectorName)"`
+}
