+++ twingate/internal/client/query/group-create.go
@@ -1,0 +1,20 @@
+package query
+
+import "github.com/Twingate/terraform-provider-twingate/twingate/internal/model"
+
+type CreateGroup struct {
+	GroupEntityResponse `graphql:"groupCreate(name: $name)"`
+}
+
+type GroupEntityResponse struct {
+	Entity *gqlGroup
+	OkError
+}
+
+func (q CreateGroup) ToModel() *model.Group {
+	if q.Entity == nil {
+		return nil
+	}
+
+	return q.Entity.ToModel()
+}
