+++ twingate/internal/client/query/group-delete.go
@@ -1,0 +1,5 @@
+package query
+
+type DeleteGroup struct {
+	OkError `graphql:"groupDelete(id: $id)" json:"groupDelete"`
+}
