+++ twingate/internal/client/query/group-update.go
@@ -1,0 +1,5 @@
+package query
+
+type UpdateGroup struct {
+	GroupEntityResponse `graphql:"groupUpdate(id: $id, name: $name)"`
+}
