+++ twingate/internal/client/query/groups-by-name-read.go
@@ -1,0 +1,5 @@
+package query
+
+type ReadGroupsByName struct {
+	Groups `graphql:"groups(filter: {name: {eq: $name}}, after: $groupsEndCursor)"`
+}
