+++ twingate/internal/client/query/pagination.go
@@ -1,4 +1,4 @@
-package client
+package query
 
 import (
 	"context"
@@ -6,14 +6,19 @@ import (
 	"github.com/twingate/go-graphql-client"
 )
 
+type PageInfo struct {
+	EndCursor   graphql.String
+	HasNextPage graphql.Boolean
+}
+
 type PaginatedResource[E any] struct {
 	PageInfo PageInfo
 	Edges    []E
 }
 
-type nextPageFunc[E any] func(ctx context.Context, variables map[string]interface{}, cursor graphql.String) (*PaginatedResource[E], error)
+type NextPageFunc[E any] func(ctx context.Context, variables map[string]interface{}, cursor graphql.String) (*PaginatedResource[E], error)
 
-func (r *PaginatedResource[E]) fetchPages(ctx context.Context, fetchNextPage nextPageFunc[E], variables map[string]interface{}) error {
+func (r *PaginatedResource[E]) FetchPages(ctx context.Context, fetchNextPage NextPageFunc[E], variables map[string]interface{}) error {
 	if r == nil {
 		return nil
 	}
