+++ twingate/internal/client/query/remote-network-by-name-read.go
@@ -1,0 +1,9 @@
+package query
+
+type ReadRemoteNetworkByName struct {
+	RemoteNetworks gqlRemoteNetworks `graphql:"remoteNetworks(filter: {name: {eq: $name}})"`
+}
+
+type gqlRemoteNetworks struct {
+	Edges []*RemoteNetworkEdge
+}
