+++ twingate/internal/client/query/remote-network-delete.go
@@ -1,0 +1,5 @@
+package query
+
+type DeleteRemoteNetwork struct {
+	OkError `graphql:"remoteNetworkDelete(id: $id)"`
+}
