+++ twingate/internal/client/query/remote-networks-read.go
@@ -1,0 +1,26 @@
+package query
+
+import (
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/model"
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/utils"
+)
+
+const CursorRemoteNetworks = "remoteNetworksEndCursor"
+
+type ReadRemoteNetworks struct {
+	RemoteNetworks `graphql:"remoteNetworks(after: $remoteNetworksEndCursor)"`
+}
+
+type RemoteNetworks struct {
+	PaginatedResource[*RemoteNetworkEdge]
+}
+
+type RemoteNetworkEdge struct {
+	Node gqlRemoteNetwork
+}
+
+func (r RemoteNetworks) ToModel() []*model.RemoteNetwork {
+	return utils.Map[*RemoteNetworkEdge, *model.RemoteNetwork](r.Edges, func(edge *RemoteNetworkEdge) *model.RemoteNetwork {
+		return edge.Node.ToModel()
+	})
+}
