+++ twingate/internal/client/query/resource-create.go
@@ -1,0 +1,10 @@
+package query
+
+type CreateResource struct {
+	ResourceEntityResponse `graphql:"resourceCreate(name: $name, address: $address, remoteNetworkId: $remoteNetworkId, groupIds: $groupIds, protocols: $protocols)"`
+}
+
+type ResourceEntityResponse struct {
+	Entity *gqlResource
+	OkError
+}
