+++ twingate/internal/client/query/resource-delete.go
@@ -1,0 +1,5 @@
+package query
+
+type DeleteResource struct {
+	OkError `graphql:"resourceDelete(id: $id)"`
+}
