+++ twingate/internal/client/query/resource-update.go
@@ -1,0 +1,9 @@
+package query
+
+type UpdateResource struct {
+	ResourceEntityResponse `graphql:"resourceUpdate(id: $id, name: $name, address: $address, remoteNetworkId: $remoteNetworkId, groupIds: $groupIds, protocols: $protocols)"`
+}
+
+type UpdateResourceActiveState struct {
+	OkError `graphql:"resourceUpdate(id: $id, isActive: $isActive)"`
+}
