+++ twingate/internal/client/query/resources-by-name-read.go
@@ -1,0 +1,5 @@
+package query
+
+type ReadResourcesByName struct {
+	Resources `graphql:"resources(filter: {name: {eq: $name}}, after: $resourcesEndCursor)"`
+}
