+++ twingate/internal/client/query/service-account-delete.go
@@ -1,0 +1,5 @@
+package query
+
+type DeleteServiceAccount struct {
+	OkError `graphql:"serviceAccountDelete(id: $id)"`
+}
