+++ twingate/internal/client/query/users-read.go
@@ -1,0 +1,26 @@
+package query
+
+import (
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/model"
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/utils"
+)
+
+const CursorUsers = "usersEndCursor"
+
+type ReadUsers struct {
+	Users `graphql:"users(after: $usersEndCursor)"`
+}
+
+type Users struct {
+	PaginatedResource[*UserEdge]
+}
+
+type UserEdge struct {
+	Node *gqlUser
+}
+
+func (u Users) ToModel() []*model.User {
+	return utils.Map[*UserEdge, *model.User](u.Edges, func(edge *UserEdge) *model.User {
+		return edge.Node.ToModel()
+	})
+}
