+++ twingate/internal/client/user.go
@@ -3,78 +3,48 @@ package client
 import (
 	"context"
 
+	"github.com/Twingate/terraform-provider-twingate/twingate/internal/client/query"
 	"github.com/Twingate/terraform-provider-twingate/twingate/internal/model"
 	"github.com/twingate/go-graphql-client"
 )
 
 const userResourceName = "user"
 
-type gqlUser struct {
-	ID        graphql.ID
-	FirstName graphql.String
-	LastName  graphql.String
-	Email     graphql.String
-	Role      graphql.String
-}
-
-type UserEdge struct {
-	Node *gqlUser
-}
-
-type Users struct {
-	PaginatedResource[*UserEdge]
-}
-
-type readUsersQuery struct {
-	Users Users
-}
-
 func (client *Client) ReadUsers(ctx context.Context) ([]*model.User, error) {
-	response := readUsersQuery{}
-
-	err := client.GraphqlClient.NamedQuery(ctx, "readUsers", &response, nil)
-	if err != nil {
-		return nil, NewAPIErrorWithID(err, "read", userResourceName, "All")
-	}
-
-	if len(response.Users.Edges) == 0 {
-		return nil, nil
-	}
-
-	err = response.Users.fetchPages(ctx, client.readUsersAfter, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return response.Users.ToModel(), nil
-}
-
-type readUsersAfter struct {
-	Users Users `graphql:"users(after: $usersEndCursor)"`
-}
-
-func (client *Client) readUsersAfter(ctx context.Context, variables map[string]interface{}, cursor graphql.String) (*PaginatedResource[*UserEdge], error) {
-	if variables == nil {
-		variables = make(map[string]interface{})
-	}
-
-	variables["usersEndCursor"] = cursor
-	response := readUsersAfter{}
+	variables := newVars(gqlNullable("", query.CursorUsers))
+	response := query.ReadUsers{}
 
 	err := client.GraphqlClient.NamedQuery(ctx, "readUsers", &response, variables)
 	if err != nil {
 		return nil, NewAPIErrorWithID(err, "read", userResourceName, "All")
 	}
 
-	if len(response.Users.Edges) == 0 {
+	if len(response.Edges) == 0 {
+		return nil, nil
+	}
+
+	err = response.FetchPages(ctx, client.readUsersAfter, variables)
+	if err != nil {
+		return nil, err //nolint
+	}
+
+	return response.ToModel(), nil
+}
+
+func (client *Client) readUsersAfter(ctx context.Context, variables map[string]interface{}, cursor graphql.String) (*query.PaginatedResource[*query.UserEdge], error) {
+	variables[query.CursorUsers] = cursor
+	response := query.ReadUsers{}
+
+	err := client.GraphqlClient.NamedQuery(ctx, "readUsers", &response, variables)
+	if err != nil {
+		return nil, NewAPIErrorWithID(err, "read", userResourceName, "All")
+	}
+
+	if len(response.Edges) == 0 {
 		return nil, NewAPIErrorWithID(ErrGraphqlResultIsEmpty, "read", userResourceName, "All")
 	}
 
-	return &response.Users.PaginatedResource, nil
-}
-
-type readUserQuery struct {
-	User *gqlUser `graphql:"user(id: $id)"`
+	return &response.PaginatedResource, nil
 }
 
 func (client *Client) ReadUser(ctx context.Context, userID string) (*model.User, error) {
@@ -83,7 +53,7 @@ func (client *Client) ReadUser(ctx context.Context, userID string) (*model.User,
 	}
 
 	variables := newVars(gqlID(userID))
-	response := readUserQuery{}
+	response := query.ReadUser{}
 
 	err := client.GraphqlClient.NamedQuery(ctx, "readUser", &response, variables)
 	if err != nil {
@@ -94,5 +64,5 @@ func (client *Client) ReadUser(ctx context.Context, userID string) (*model.User,
 		return nil, NewAPIErrorWithID(ErrGraphqlResultIsEmpty, "read", userResourceName, userID)
 	}
 
-	return response.User.ToModel(), nil
+	return response.ToModel(), nil
 }
