+++ twingate/internal/test/models/resource_test.go
@@ -82,10 +82,10 @@ func TestNewPortRange(t *testing.T) {
 }
 
 func TestResourceModel(t *testing.T) {
-	var emptySlice []interface{}
-	var emptyStringSlice []string
-	{
-	}
+	var (
+		emptySlice       []interface{}
+		emptyStringSlice []string
+	)
 
 	cases := []struct {
 		resource model.Resource
@@ -195,4 +195,23 @@ func TestNewProtocol(t *testing.T) {
 			assert.Equal(t, c.expected, model.NewProtocol(c.policy, c.ports))
 		})
 	}
+}
+
+func TestProtocols(t *testing.T) {
+	t.Run("Test Twingate Resource : Protocols", func(t *testing.T) {
+		protocols := model.DefaultProtocols()
+
+		assert.EqualValues(t, model.PolicyAllowAll, protocols.TCP.Policy)
+		assert.EqualValues(t, model.PolicyAllowAll, protocols.UDP.Policy)
+		assert.Nil(t, protocols.UDP.Ports)
+		assert.Nil(t, protocols.TCP.Ports)
+
+		port := &model.PortRange{Start: 1, End: 18000}
+		protocols.TCP.Ports = append(protocols.TCP.Ports, port)
+		protocols.UDP.Ports = append(protocols.UDP.Ports, port)
+		udpPorts := protocols.UDP.PortsToString()
+		tcpPorts := protocols.TCP.PortsToString()
+		assert.EqualValues(t, "1-18000", tcpPorts[0])
+		assert.EqualValues(t, "1-18000", udpPorts[0])
+	})
 }
